New test for validations and iiif json harvester, refactor csv harvester test
